Create placeholder.txt
The "data/" folder contains newly created CSV files like the list of .Rdm file paths to be translated (data/filesToTranslate.csv), the list of languages to choose for translation (data/languages.csv), the list of excluded files for translation in a particular language.
The filesToTranslate.csv contains all the paths of the files in the source version of the OHDSI book. They should all be present in the translated releases.  The Rmd files in the book should all be translated and therefore their paths are in the filesToTranslate.csv. But not all the image files can have their text translated. Therefore, only the images that can/must be translated are in the  filesToTranslate.csv. Image (or other) files -new or existing in the source book-  can be added to this list anytime. The way to treat their translation will be worked out and defined later. The idea is to make translations dynamic so that every translated book can easily be synced to the original one. But maybe some illustrative files are difficult to translate in a particular language or should be replaced by another illustration. Therefore we have the list "exception_files" whith three columns: 1/ "filePath", 2/ "language",  3/ exception ("not to translate"/"extra"/others...?).